Fix x64 build.

+++ RemoteFileBrowser/RemoteFileBrowser/Utilities/StreamableFile.h
@@ -11,7 +11,7 @@ public:
 	StreamableFile(const std::wstring& filePath);
 	~StreamableFile();
 
-	static const int kMaxChunkSize = 8 * 1024 * 1024; // 8 MB at a time
+	static const size_t kMaxChunkSize = 8 * 1024 * 1024; // 8 MB at a time
 
 	bool IsEndOfFile() const { return m_FilePosition == m_FileSize; }
 	inline uint64_t GetFileSize() const { return m_FileSize; }
